own changes

+++ Backend/Dto/ProductDto.cs
@@ -5,5 +5,5 @@ public class ProductDto
     public string Naam { get; set; } = null!;
     public string ArtikelKenmerken { get; set; } = null!;
     public int Hoeveelheid { get; set; }
-    public float StartPrijs { get; set; }
+    public decimal StartPrijs { get; set; }
 }
