+++ backend/Dto/ProductCreateDto.cs
@@ -4,6 +4,6 @@ public class ProductCreateDto
     public string Naam { get; set; } = null!;
     public string ArtikelKenmerken { get; set; } = null!;
     public int Hoeveelheid { get; set; }
-    public float StartPrijs { get; set; }
+    public decimal StartPrijs { get; set; }
     public string Aanvoerder_Id { get; set; } = null!;
 }
